Allow users to pass new properties on restore
VM changes: To identify properties that can be modified on restore,
added a new bool field SystemProperty::_modifiable_on_restore.
All the jdk related properties are marked unmodifiable. Rest of the
properties are considered modifiable.
When the JVM is launched with -XX:CRaCRestoreFrom option, then the
properties prefixed with "-D" are maintained in a separate list in
Arguments::_system_properties_for_restore. This list is passed to the JVM
being restored by writing to a shared memory object.
When the JVM is restored, it reads the new properties from shared memory
object and updates its existing list of properties maintained in
Arguments::_system_properties.

JDK changes: System::props needs to be updated on restore to account for
new system properties. For this purpose j.l.System registers a new
JDKResource which queries new properties from the VM in afterRestore()
notification and updates System::props. The JDKResource registered by
j.l.System is given highest priority so it is the first resource to get
afterRestore() notification.

Signed-off-by: Ashutosh Mehra <asmehra@redhat.com>